add to utils format string case
