+++ packages/next-app/src/contexts/AdminToolsContext.tsx
@@ -9,26 +9,6 @@ import {
   useState,
 } from "react";
 
-export function toSlug(string?: string) {
-  return (
-    string
-      ?.toLowerCase()
-      ?.replace(/ /g, "-")
-      ?.replace(/[^\w-]+/g, "") || ""
-  );
-}
-
-export function toCammelCase(string: string) {
-  return string
-    .replace(/\s(.)/g, function ($1) {
-      return $1.toUpperCase();
-    })
-    .replace(/\s/g, "")
-    .replace(/^(.)/, function ($1) {
-      return $1.toLowerCase();
-    });
-}
-
 // TODO: generate TS types from zod: https://github.com/sachinraja/zod-to-ts
 export interface ActionAttribute {
   id: number;
