+++ packages/next-app/src/utils/formatStringCase.ts
@@ -1,0 +1,19 @@
+export function toSlug(string?: string) {
+    return (
+      string
+        ?.toLowerCase()
+        ?.replace(/ /g, "-")
+        ?.replace(/[^\w-]+/g, "") || ""
+    );
+  }
+  
+export function toCamelCase(string: string) {
+    return string
+      .replace(/\s(.)/g, function ($1) {
+        return $1.toUpperCase();
+      })
+      .replace(/\s/g, "")
+      .replace(/^(.)/, function ($1) {
+        return $1.toLowerCase();
+      });
+}
